@@ -45,19 +45,18 @@
   "devDependencies": {
     "babel-eslint": "^7.1.1",
     "babel-jest": "^17.0.2",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
     "eslint": "^3.10.2",
-    "eslint-plugin-react": "^6.7.1",
+    "eslint-plugin-react": "^6.10.3",
     "jest": "^17.0.3",
     "now": "^4.0.1"
   },
   "dependencies": {
     "lodash": "^4.17.4",
-    "next": "^2.0.1",
+    "next": "^2.1.1",
     "nova-colors": "^2.1.5",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.5.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-tabs": "^0.8.2"
   }
 }
