@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import Link from 'next/link'
 import {colorValues} from '../utils/styleGuide'
 
