@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import {uiGroups, borderRadii, borderSizes} from '../utils/styleGuide'
 
 const Image = ({src, description, quiet = false, style}) => (
